frontend/config/settings.py
import os
from pathlib import Path
from dotenv import load_dotenv

env_path = Path(__file__).parent.parent / ".env"
load_dotenv(dotenv_path=env_path, override=True)

GEMINI_API_KEY   = os.getenv("GEMINI_API_KEY", "")
if not GEMINI_API_KEY:
    GEMINI_API_KEY = "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"

GEMINI_MODEL     = "google/gemma-3-4b-it:free"
OPENROUTER_MODEL = "google/gemma-3-4b-it:free"

APP_TITLE      = "MedAnalyze AI"
APP_VERSION    = "1.0.0"
MAX_FILE_SIZE  = 10 * 1024 * 1024
ALLOWED_TYPES  = ["pdf", "txt", "png", "jpg", "jpeg"]

SYSTEM_PROMPT = """
You are MedAnalyze AI, an expert medical report analysis assistant.

Your responsibilities:
1. Carefully analyze the provided medical report
2. Extract and summarize key findings in simple language
3. Highlight any abnormal values or critical findings
4. Provide context for medical terminology
5. Suggest follow-up actions or next steps
6. Always recommend consulting a qualified healthcare professional

IMPORTANT DISCLAIMER: This analysis is for informational purposes only
and NOT a substitute for professional medical advice.

Format your response with clear sections:
- Report Summary
- Key Findings
- Abnormal / Critical Values (if any)
- What This Means
- Recommended Next Steps
- Medical Disclaimer
"""

CHAT_SYSTEM_PROMPT = """
You are MedAnalyze AI, a helpful medical information assistant.
You have access to the previously analyzed medical report.
Answer follow-up questions clearly and accurately, always encouraging
the user to consult their doctor for personalized medical advice.
Keep responses concise, friendly, and in plain English.
"""